♻️ Refactor sent NFT Book logic
